Updates based on feedback from Colette

+++ services/Models/Data/FishScales_Detail.cs
@@ -9,11 +9,11 @@ namespace services.Models.Data
         public DateTime? ScaleCollectionDate { get; set; }
         public string Species { get; set; }
         public string LifeStage { get; set; }
-        public int? Circuli { get; set; }
+        public double? Circuli { get; set; }
         public string JuvenileAge { get; set; }
-        public int? FreshwaterAge { get; set; }
-        public int? SaltWaterAge { get; set; }
-        public int? TotalAdultAge { get; set; }
+        public double? FreshwaterAge { get; set; }
+        public double? SaltWaterAge { get; set; }
+        public double? TotalAdultAge { get; set; }
         public string SpawnCheck { get; set; }
         public string Regeneration { get; set; }
         public string Stock { get; set; }
